:recycle: Move routes around for my sanity.
